added timeline component
