+++ client/src/Components/Timeline/timeline.css
@@ -1,0 +1,25 @@
+:import('../../colors.css') {
+  gridBlue: gridBlue;
+  blue: blue;
+}
+
+.Timeline {
+  margin-top: 10px;
+  border: 1px solid gridBlue;
+  position: relative;
+  width: 95%;
+  height: 3px;
+}
+
+.Marker {
+  border: 3px solid blue;
+  background-color: blue;
+  position: absolute;
+  box-shadow: lightestShadow;
+  left: -4px;
+  top: -8px;
+  height: 16px;
+  width: 16px;
+  border-radius: 50%;
+  cursor: pointer;
+}
